web_server/lsaas2/challenge/app/requirements.txt
flask
python-dotenv
uwsgi
subprocess.run
